fix: contract test
